fix float type
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@2508 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/army/army_troop.cpp
@@ -199,7 +199,7 @@ u32 Army::Troop::GetDamageMax(void) const
 
 u32 Army::Troop::GetStrength(void) const
 {
-    double res = (GetDamageMin() + GetDamageMax()) >> 1;
+    float res = (GetDamageMin() + GetDamageMax()) >> 1;
 
     // increase strength
     if(isFly())		res += res * 0.5;
